\name{get.stats.mcmc}
\alias{get.stats.mcmc}
\title{Get MCMC stats}
\usage{
  get.stats.mcmc(mcmc.summary, sample.size)
}
\arguments{
  \item{mcmc.summary}{}

  \item{sample.size}{}
}
\value{
  summary statistics from the results of a meta analysis
}
\description{
  Further summarizes output from summary.mcmc
}
\author{
  David LeBauer
}

