\name{theme_border}
\alias{theme_border}
\title{panel.border}
\usage{
  theme_border(type = c("left", "right", "bottom", "top", "none"),
    colour = "black", size = 1, linetype = 1)
}
\arguments{
  \item{type}{}

  \item{colour}{}

  \item{size}{}

  \item{linetype}{}
}
\value{
  a plot border
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Rudolf Cardinal http://goo.gl/YThRT
}

